+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented here.
 ## v0.8.0 - 2022-12-30
 ### Added
 - Some refactoring and QoL improvements to ease the development of custom [IEffectRenderer](src/main/java/de/teragam/jfxshader/IEffectRenderer.java).
+- The generated texture coordinates for the shader input textures can now be accessed in the [ShaderEffectPeer](src/main/java/de/teragam/jfxshader/ShaderEffectPeer.java). This allows the shaders to compensate for the varying texture coordinates caused by JavaFX's internal texture pooling.
+- The `ImagePoolPolicy.QUANTIZED` option has been added to allow better VRAM usage while ensuring that the target texture dimensions do not fluctuate between frames.
+- Added option to invert the mask in the [BlendShapes](src/main/java/de/teragam/jfxshader/samples/blendshapes/BlendShapesEffectPeer.java) example.
 
 ### Fixed
 - Fixed an issue that did not allow the use of all 4 textures in the shaders.